+++ src/Table.h
@@ -1,28 +1,26 @@
 #ifndef TABLE_H
 #define TABLE_H
 
-#include <vector>
-
 #include "Column.h"
 #include "MemoryAllocator.h"
 
 class Table
 {
     MemoryAllocator memory_allocator; // Memory Allocator to generate new chunk
-    int attribute_count;              // Number of int columns
-    int value_count;                  // Number of double columns
+    unsigned int attribute_count;     // Number of int columns
+    unsigned int value_count;         // Number of double columns
     std::vector<Column> columns;      // List of table's column
 
 public:
-    Table(int attribute_count, int value_count);
+    Table(unsigned int attribute_count, unsigned int value_count);
     ~Table() = default;
 
-    int readInt(int row_pos, int column);
-    double readDouble(int row_pos, int column);
-    void writeInt(int row_pos, int column, int value);
-    void writeDouble(int row_pos, int column, double value);
-    void writeRow(int row_pos, int *attributes, double *values);
-    void writeRows(int starting_row_pos, int number_of_rows, int *attributes, double *values);
+    int readInt(unsigned long int row_pos, unsigned int column);
+    double readDouble(unsigned long int row_pos, unsigned int column);
+    void writeInt(unsigned long int row_pos, unsigned int column, unsigned int value);
+    void writeDouble(unsigned long int row_pos, unsigned int column, double value);
+    void writeRow(unsigned long int row_pos, unsigned int *attributes, double *values);
+    void writeRows(unsigned long int starting_row_pos, unsigned long int number_of_rows, unsigned int *attributes, double *values);
 
     std::vector<std::bitset<CHUNK_SIZE>> findRows(int *attribute_predicates, double *value_predicates);
 };
